move nml editor dialog into separate widget modules
