+++ gis4wrf/plugin/ui/widget_run.py
@@ -32,6 +32,8 @@ from gis4wrf.core import Project, read_namelist, verify_namelist, get_namelist_s
 from gis4wrf.plugin.constants import PLUGIN_NAME
 from gis4wrf.plugin.options import get_options
 from gis4wrf.plugin.ui.helpers import MessageBar, IgnoreKeyPressesDialog
+from gis4wrf.plugin.ui.browser_nml_schema import NmlSchemaBrowser
+from gis4wrf.plugin.ui.dialog_nml_editor import NmlEditorDialog
 
 class RunWidget(QWidget):
     tab_active = pyqtSignal()
@@ -98,16 +100,6 @@ class RunWidget(QWidget):
         ''' Sets the currently active project. See tab_simulation. '''
         self._project = val
 
-    def verify_wps_nml(self, text: str) -> None:
-        nml = read_namelist(StringIO(text))
-        # TODO catch "'ascii' codec can't decode byte 0xe2 in position 200: ordinal not in range(128)"
-        #      and either auto-convert to ascii or display more helpful message
-        verify_namelist(nml, 'wps')
-
-    def verify_wrf_nml(self, text: str) -> None:
-        nml = read_namelist(StringIO(text))
-        verify_namelist(nml, 'wrf')
-
     def prepare_wps_run(self) -> None:
         self.project.prepare_wps_run(self.options.wps_dir)
 
@@ -117,7 +109,6 @@ class RunWidget(QWidget):
     def on_open_namelist_wps_clicked(self) -> None:
         self.project.update_wps_namelist()
         self.open_editor_dialog(self.project.wps_namelist_path,
-                                self.verify_wps_nml,
                                 get_namelist_schema('wps'))   
 
     def on_open_namelist_wrf_clicked(self) -> None:
@@ -125,7 +116,6 @@ class RunWidget(QWidget):
         #      (same for prepare_*_run)
         self.project.update_wrf_namelist()
         self.open_editor_dialog(self.project.wrf_namelist_path,
-                                self.verify_wrf_nml,
                                 get_namelist_schema('wrf'))
 
     def on_prepare_only_wps_clicked(self) -> None:
@@ -216,176 +206,9 @@ class RunWidget(QWidget):
             QMessageBox.information(self.iface.mainWindow(), PLUGIN_NAME, 
                 'Program {} finished without errors!'.format(os.path.basename(path)))
 
-    def open_editor_dialog(self, path: str, validate: Callable[[str],Optional[str]],
-                           nml_schema: dict) -> Optional[str]:
-        with open(path, 'r') as fp:
-            text = fp.read()
-
-        geom = QGuiApplication.primaryScreen().geometry()
-        w, h = geom.width(), geom.height()
-        
-        # Ignoring key presses is necessary as otherwise pressing
-        # enter in the search box closes the dialog.
-        dialog = IgnoreKeyPressesDialog()
-        
-        dialog.setWindowTitle(os.path.basename(path))
-        dialog.resize(w * 0.7, h * 0.6)
-        
-        layout = QVBoxLayout()
-        dialog.setLayout(layout)
-
-        hbox = QHBoxLayout()
-        layout.addLayout(hbox)
-
-        editor = QPlainTextEdit(text, dialog)
-        editor.setWordWrapMode(QTextOption.NoWrap)
-        hbox.addWidget(editor)
-
-        def to_fortran(val) -> str:
-            if val is True:
-                return '.true.'
-            if val is False:
-                return '.false.'
-            if isinstance(val, str):
-                val = f"'{val}'"
-            return str(val)
-
-        nml_help = '<html>'
-        anchors = set() # type: Set[str]
-        for section_name, section in nml_schema.items():
-            anchors.add(section_name)
-            nml_help += f'<h2><a name="{section_name}">&{section_name}</a></h2>'
-            for var_name, variable in section.items():
-                anchors.add(var_name)
-                description = variable['description']
-                type_ = variable['type']
-                item_type = variable.get('itemtype')
-                min_len = variable.get('minlen')
-                min_ = variable.get('min')
-                max_ = variable.get('max')
-                if item_type:
-                    type_label += f' of {item_type}'
-                else:
-                    type_label = type_
-                default = variable.get('default')
-                example = variable.get('example')
-                options = variable.get('options')
-
-                nml_help += f'<h3><a name="{var_name}">{var_name}</a></h3>'
-                nml_help += f'<p>{description}</p>'
-                nml_help += f'Type: {type_}<br>'
-                if min_len is not None:
-                    if isinstance(min_len, str):
-                        min_len = f'<a href="#{min_len}">{min_len}</a>'
-                    nml_help += f'Length: {min_len}<br>'
-                if min_ is not None:
-                    nml_help += f'Min: <code>{min_}</code><br>'
-                if max_ is not None:
-                    nml_help += f'Max: <code>{max_}</code><br>'
-                if default is not None:
-                    default_ = to_fortran(default)
-                    if type_ == 'list':
-                        if default == []:
-                            nml_help += f'Default: empty list'
-                        else:
-                            nml_help += f'Default: list of <code>{default_}</code>'
-                    else:
-                        nml_help += f'Default: <code>{default_}</code><br>'
-                if example is not None:
-                    val_type = item_type if item_type else type_
-                    if isinstance(example, str) and val_type != 'str':
-                        # Here we have a literal example in Fortran syntax,
-                        # so we avoid surrounding it with single quotes.
-                        pass
-                    else:
-                        example = to_fortran(example)
-                    nml_help += f'Example: <code>{example}</code><br>'
-                if options:
-                    if isinstance(options, list):
-                        nml_help += 'Options: <code>' + ', '.join(map(to_fortran, options)) + '</code><br>'
-                    else:
-                        nml_help += '<br>Options: <table border=1>'
-                        for val, description in options.items():
-                            val = to_fortran(val)
-                            nml_help += f'<tr><td width="30%" align="center"><code>{val}</code></td>'
-                            nml_help += f'<td width="70%">{description}</td></tr>'
-                        nml_help += '</table>'
-
-        nml_help += '</html>'
-
-        nml_help_vbox = QVBoxLayout()
-        hbox.addLayout(nml_help_vbox)
-
-        # TODO set encoding of help page, e.g. degree symbol appears incorrect
-        nml_help_viewer = QTextBrowser()
-        nml_help_viewer.setMaximumWidth(dialog.width() * 0.3)
-        nml_help_viewer.setText(nml_help)
-
-        nml_help_search = QLineEdit()
-        nml_help_search.setMaximumWidth(nml_help_viewer.maximumWidth())
-        nml_help_search.setPlaceholderText('Find...')
-
-        def search_from_start():
-            nml_help_viewer.moveCursor(QTextCursor.Start)
-            nml_help_viewer.find(nml_help_search.text())
-
-            # TODO scroll such that search term is vertically centered
-            #      currently page is minimally scrolled and term is often at the bottom
-        
-        def search_from_cursor():
-            found = nml_help_viewer.find(nml_help_search.text())
-            if not found:
-                search_from_start()
-
-        nml_help_search.textChanged.connect(search_from_start)
-        nml_help_search.returnPressed.connect(search_from_cursor)
-
-        nml_help_vbox.addWidget(nml_help_search)
-        nml_help_vbox.addWidget(nml_help_viewer)
-
-        def on_cursor_pos_changed():
-            cursor = editor.textCursor()
-            line = cursor.block().text()
-            match = re.match(r'\s*&?([a-zA-Z0-9_]+)', line)
-            if not match:
-                return
-            name = match.group(1)
-            if name not in anchors:
-                return
-            nml_help_viewer.scrollToAnchor(name)
-
-        editor.cursorPositionChanged.connect(on_cursor_pos_changed)
-
-        def validate_and_save():
-            text = editor.toPlainText()
-            try:
-                validate(text)
-            except (TypeError, ValueError) as e:
-                btn = QMessageBox.critical(
-                    dialog, 'Syntax error', str(e),
-                    QMessageBox.Cancel | QMessageBox.Ignore,
-                    QMessageBox.Cancel)
-                if btn != QMessageBox.Ignore:
-                    return
-            with open(path, 'w') as fp:
-                fp.write(text)
-            dialog.accept()
-
-        button_box = QDialogButtonBox(QDialogButtonBox.Save | QDialogButtonBox.Cancel)
-        button_box.accepted.connect(validate_and_save)
-        button_box.rejected.connect(dialog.reject)
-        layout.addWidget(button_box)
-
-        doc = editor.document()
-        font = doc.defaultFont()
-        font.setFamily("Courier New")
-        font.setPointSize(11)
-        doc.setDefaultFont(font)
-
-        result = dialog.exec_()
-        if result == QDialog.Rejected:
-            return None
-        return editor.toPlainText()
+    def open_editor_dialog(self, path: str, nml_schema: dict) -> Optional[str]:
+        dialog = NmlEditorDialog(path, nml_schema)
+        dialog.exec_()
 
     def create_gbox(self, gbox_name: str, btn_names: List[Union[str,List[str]]]) \
             -> Tuple[QGroupBox, List[QPushButton]]:
